Remove irrelevant comment from transaction.rs
The comment about "comparing C-like enums" refers to the use of
(state as int) == (Done as int), which is no longer used since
a minor refactor to fail on partially-decoded transactions.

I still don't know what the correct way to compare enums is. :)
Doing state == Done gives a compiler error because Done is a type,
not an rvalue.