Add some use statement to handle changes in Rust's module searching
I need to add 'use hash::Hashable' in order to use transaction.to_hash(),
as to_hash() is part of the hash::Hashable trait. (Before, using
transaction itself would pull in to_hash().)

This gives an 'unnecessary use statement' warning, which suggests
that this is not the proper fix. I guess a TODO is to go on the
Rust IRC and ask them what's up.