Add simple test suite
It's a start, anyway. Right now this can't handle unsigned merges
since those are randomize, and it doesn't really handle error
reporting properly since it only looks at the last 'err:' line.
Also it is very incomplete.

This is much easier to use than the old method -- manually running
a bunch of test-* files and checking them with decoderawtransaction.